changed module names to remove multiple versions
